renamed ACExplorer to pyUbiForge to avoid confusion
previously ACExplorer (the UI side of the program) and the back end code were very intertwined. The last number of commits have seperated the two such that the back end code is its own entity that can be used with other applications and the UI is just a wrapper.
To make this more obvious the back end code has been renamed to pyUbiForge
