+++ pyUbiForge/ACU/formatFile.py
@@ -4,7 +4,7 @@ import sys
 import json
 import time
 import struct
-from ACExplorer.misc.file_object import FileObject
+from pyUbiForge.misc.file_object import FileObject
 # from ACExplorer.misc.dataTypes import LE2BE2, BEHEX2, LE2DEC2, int16, uint16, int32, uint32, float32
 fileTypes = json.load(open(r"./ACExplorer/ACU/fileFormats.json"))
 indentCharacter = '\t'
