+++ pyUbiForge/misc/file_object.py
