+++ pyUbiForge/misc/texture.py
